# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CategoryGender =
  (typeof CategoryGender)[keyof typeof CategoryGender];

export const CategoryGender = {
  men: "men",
  women: "women",
  kids: "kids",
  unisex: "unisex",
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  gender?: CategoryGender;
}

export type ProductGender = (typeof ProductGender)[keyof typeof ProductGender];

export const ProductGender = {
  men: "men",
  women: "women",
  kids: "kids",
  unisex: "unisex",
} as const;

export type ProductAgeGroup =
  (typeof ProductAgeGroup)[keyof typeof ProductAgeGroup];

export const ProductAgeGroup = {
  adults: "adults",
  kids: "kids",
  all: "all",
} as const;

export type ProductBadge = (typeof ProductBadge)[keyof typeof ProductBadge];

export const ProductBadge = {
  new: "new",
  hot: "hot",
  sale: "sale",
} as const;

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  categoryId: number;
  categoryName?: string;
  categorySlug?: string;
  gender?: ProductGender;
  ageGroup?: ProductAgeGroup;
  sizes?: string[];
  colors?: string[];
  material?: string;
  imageUrls?: string[];
  stockQuantity: number;
  lowStockThreshold?: number;
  isAvailable: boolean;
  isFeatured?: boolean;
  badge?: ProductBadge;
  createdAt?: string;
}

export type CreateProductRequestGender =
  (typeof CreateProductRequestGender)[keyof typeof CreateProductRequestGender];

export const CreateProductRequestGender = {
  men: "men",
  women: "women",
  kids: "kids",
  unisex: "unisex",
} as const;

export type CreateProductRequestAgeGroup =
  (typeof CreateProductRequestAgeGroup)[keyof typeof CreateProductRequestAgeGroup];

export const CreateProductRequestAgeGroup = {
  adults: "adults",
  kids: "kids",
  all: "all",
} as const;

export type CreateProductRequestBadge =
  (typeof CreateProductRequestBadge)[keyof typeof CreateProductRequestBadge];

export const CreateProductRequestBadge = {
  new: "new",
  hot: "hot",
  sale: "sale",
} as const;

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  categoryId: number;
  gender?: CreateProductRequestGender;
  ageGroup?: CreateProductRequestAgeGroup;
  sizes?: string[];
  colors?: string[];
  material?: string;
  imageUrls?: string[];
  stockQuantity: number;
  lowStockThreshold?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  badge?: CreateProductRequestBadge;
}

export type UpdateProductRequestGender =
  (typeof UpdateProductRequestGender)[keyof typeof UpdateProductRequestGender];

export const UpdateProductRequestGender = {
  men: "men",
  women: "women",
  kids: "kids",
  unisex: "unisex",
} as const;

export type UpdateProductRequestAgeGroup =
  (typeof UpdateProductRequestAgeGroup)[keyof typeof UpdateProductRequestAgeGroup];

export const UpdateProductRequestAgeGroup = {
  adults: "adults",
  kids: "kids",
  all: "all",
} as const;

export type UpdateProductRequestBadge =
  (typeof UpdateProductRequestBadge)[keyof typeof UpdateProductRequestBadge];

export const UpdateProductRequestBadge = {
  new: "new",
  hot: "hot",
  sale: "sale",
} as const;

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  categoryId?: number;
  gender?: UpdateProductRequestGender;
  ageGroup?: UpdateProductRequestAgeGroup;
  sizes?: string[];
  colors?: string[];
  material?: string;
  imageUrls?: string[];
  stockQuantity?: number;
  lowStockThreshold?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  badge?: UpdateProductRequestBadge;
}

export interface AdminLoginRequest {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  message: string;
}

export interface UploadImageResponse {
  url: string;
  filename: string;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  inStock?: boolean;
};

export type UploadImageBody = {
  image: Blob;
};