# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Basmallah Fashion Sales API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: categories
    description: Category operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: inStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials

  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin list all products including out-of-stock
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /admin/products/{id}:
    put:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  /admin/upload:
    post:
      operationId: uploadImage
      tags: [admin]
      summary: Upload a product image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
                - image
      responses:
        "200":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImageResponse"

  /admin/inventory/alerts:
    get:
      operationId: getInventoryAlerts
      tags: [admin]
      summary: Get low stock alerts
      responses:
        "200":
          description: Low stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        gender:
          type: string
          enum: [men, women, kids, unisex]
      required:
        - id
        - name
        - slug

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        gender:
          type: string
          enum: [men, women, kids, unisex]
        ageGroup:
          type: string
          enum: [adults, kids, all]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        badge:
          type: string
          enum: [new, hot, sale]
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - categoryId
        - stockQuantity
        - isAvailable

    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        gender:
          type: string
          enum: [men, women, kids, unisex]
        ageGroup:
          type: string
          enum: [adults, kids, all]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        badge:
          type: string
          enum: [new, hot, sale]
      required:
        - name
        - price
        - categoryId
        - stockQuantity

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        gender:
          type: string
          enum: [men, women, kids, unisex]
        ageGroup:
          type: string
          enum: [adults, kids, all]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        badge:
          type: string
          enum: [new, hot, sale]

    AdminLoginRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        message:
          type: string
      required:
        - token
        - message

    UploadImageResponse:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
      required:
        - url
        - filename