# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  description: zod.string().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List products
 */
export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  inStock: zod.coerce.boolean().optional(),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  categorySlug: zod.string().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
  createdAt: zod.string().optional(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Get a product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  categorySlug: zod.string().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  message: zod.string(),
});

/**
 * @summary Admin list all products including out-of-stock
 */
export const AdminListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  categorySlug: zod.string().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
  createdAt: zod.string().optional(),
});
export const AdminListProductsResponse = zod.array(
  AdminListProductsResponseItem,
);

/**
 * @summary Create a product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  categoryId: zod.number(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean().optional(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  price: zod.number().optional(),
  categoryId: zod.number().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number().optional(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean().optional(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  categorySlug: zod.string().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Upload a product image
 */
export const UploadImageBody = zod.object({
  image: zod.instanceof(File),
});

export const UploadImageResponse = zod.object({
  url: zod.string(),
  filename: zod.string(),
});

/**
 * @summary Get low stock alerts
 */
export const GetInventoryAlertsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  categorySlug: zod.string().optional(),
  gender: zod.enum(["men", "women", "kids", "unisex"]).optional(),
  ageGroup: zod.enum(["adults", "kids", "all"]).optional(),
  sizes: zod.array(zod.string()).optional(),
  colors: zod.array(zod.string()).optional(),
  material: zod.string().optional(),
  imageUrls: zod.array(zod.string()).optional(),
  stockQuantity: zod.number(),
  lowStockThreshold: zod.number().optional(),
  isAvailable: zod.boolean(),
  isFeatured: zod.boolean().optional(),
  badge: zod.enum(["new", "hot", "sale"]).optional(),
  createdAt: zod.string().optional(),
});
export const GetInventoryAlertsResponse = zod.array(
  GetInventoryAlertsResponseItem,
);