# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryGender } from "./categoryGender";

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  gender?: CategoryGender;
}