# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductRequestAgeGroup } from "./createProductRequestAgeGroup";
import type { CreateProductRequestBadge } from "./createProductRequestBadge";
import type { CreateProductRequestGender } from "./createProductRequestGender";

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  categoryId: number;
  gender?: CreateProductRequestGender;
  ageGroup?: CreateProductRequestAgeGroup;
  sizes?: string[];
  colors?: string[];
  material?: string;
  imageUrls?: string[];
  stockQuantity: number;
  lowStockThreshold?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  badge?: CreateProductRequestBadge;
}