# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductAgeGroup } from "./productAgeGroup";
import type { ProductBadge } from "./productBadge";
import type { ProductGender } from "./productGender";

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  categoryId: number;
  categoryName?: string;
  categorySlug?: string;
  gender?: ProductGender;
  ageGroup?: ProductAgeGroup;
  sizes?: string[];
  colors?: string[];
  material?: string;
  imageUrls?: string[];
  stockQuantity: number;
  lowStockThreshold?: number;
  isAvailable: boolean;
  isFeatured?: boolean;
  badge?: ProductBadge;
  createdAt?: string;
}