# lib/api-zod/src/generated/types/updateProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProductRequestAgeGroup } from "./updateProductRequestAgeGroup";
import type { UpdateProductRequestBadge } from "./updateProductRequestBadge";
import type { UpdateProductRequestGender } from "./updateProductRequestGender";

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  categoryId?: number;
  gender?: UpdateProductRequestGender;
  ageGroup?: UpdateProductRequestAgeGroup;
  sizes?: string[];
  colors?: string[];
  material?: string;
  imageUrls?: string[];
  stockQuantity?: number;
  lowStockThreshold?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  badge?: UpdateProductRequestBadge;
}