# lib/api-zod/src/generated/types/uploadImageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmallah Fashion Sales API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadImageResponse {
  url: string;
  filename: string;
}